Rename AchievementsProvider to AchievementsRepository
