+++ src/util/achievements/index.ts
@@ -1,4 +1,4 @@
 export * from "./achievements";
 export * from "./achievement";
 export * from "./achievementOptions";
-export * from "./achievementsProvider";
+export * from "./achievementsRepository";
